# Optimize dfn.js