* 47ms was spent on creating a non-live array of document.links.
  querySelectorAll's return value is non-live already.
* 95ms was spent on .hash.substr, which I think invokes URL parser,
  serializer, and then string manipulation. Instead move the "#"
  check into the selector for querySelectorAll, and avoid invoking
  the URL parser.
* The call to .closest() is not changed but now takes 57ms instead
  of 200ms. I have no idea why.
* Avoid invoking `new Date()` 480,000 times, and instead assume
  that processing 1000 links in a single idle callback is a-ok.

All in all, the time spent in dfnInternal went down from ~750ms to
~350ms in Chrome Canary.

<img width="1658" alt="dfnjs-perf-comparison" src="https://cloud.githubusercontent.com/assets/244772/20544394/28a9df32-b10a-11e6-9ecb-3e9a6c1d9ae6.png">